+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.0-beta.3] - 2018-01-27
+### Added
+- Site Group Importing
+- Site Importing
+- Asset Volume Importing
+- Asset Transform Importing
+- Tag Group Importing
+- Category Group Importing
+- Gobal Set Importing
+- Exception & Error catching during parse and save.
+
+### Fixed
+- Fixed error in imported Asset Fields causing issue with saving sections.
+
 ## [2.0.0-beta.2] - 2018-01-27
 ### Added
 - Section Importing
@@ -16,5 +30,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Field Group Importing
 - Field Importing
 
-[Unreleased]: https://github.com/pennebaker/craft-architect/compare/2.0.0-beta.2...develop
+[Unreleased]: https://github.com/pennebaker/craft-architect/compare/2.0.0-beta.3...develop
+[2.0.0-beta.3]: https://github.com/pennebaker/craft-architect/compare/2.0.0-beta.2...2.0.0-beta.3
 [2.0.0-beta.2]: https://github.com/pennebaker/craft-architect/compare/2.0.0-beta.1...2.0.0-beta.2